backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
dj-database-url>=2.1
reportlab>=4.0
gunicorn>=21.2
